eleicoes_api/requirements.txt
django
djangorestframework
drf-yasg
qrcode
Pillow
psycopg2-binary